+++ odoo/songs/install/data_demo.py
@@ -364,6 +364,24 @@ def import_rma(ctx):
 
 
 @anthem.log
+def import_stock_inventory(ctx):
+    """ Importing stock inventory from csv """
+    model = ctx.env['stock.inventory'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/stock_inventory.csv', model)
+
+    model_line = ctx.env['stock.inventory.line'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/stock_inventory_line.csv', model_line)
+
+    inventories = ctx.env['stock.inventory'].search([])
+    for inventory in inventories:
+        inventory.action_done()
+
+
+@anthem.log
 def main(ctx):
     """ Loading demo data """
     import_users(ctx)
@@ -393,4 +411,5 @@ def main(ctx):
     import_invoices_supplier(ctx)
     import_invoices_customer(ctx)
     import_rma(ctx)
+    import_stock_inventory(ctx)
     return