+++ HISTORY.rst
@@ -28,6 +28,7 @@ latest (unreleased)
 * BSSFL-392: Add the partners properties data
 * BSSFL-396: Add Delivery methode data
 * BSSFL-107: Add routing data
+* BSSFL-108: Initial stock inventory
 
 **Features and Improvements**